requirements.txt
openai>=1.0.0
gradio>=3.35.2
python-dotenv>=1.0.0 